feat: implement HomeClient component and database migration
